AST: Split off ConformanceLookup.cpp from Module.cpp
